+++ plugins/bulkio/src/bulkio/InUInt32Port.java
@@ -361,7 +361,7 @@ public class InUInt32Port extends BULKIO.jni.dataUlongPOA {
             }
 
             synchronized (this.dataBufferLock) {
-                this.stats.update(data.length, this.workQueue.size()/this.maxQueueDepth, eos, streamID, false);
+                this.stats.update(data.length, this.workQueue.size()/(float)this.maxQueueDepth, eos, streamID, false);
                 this.workQueue.add(p);
                 this.dataSem.release();
             }
@@ -386,7 +386,7 @@ public class InUInt32Port extends BULKIO.jni.dataUlongPOA {
                     this.stats.update(data.length, 0, eos, streamID, true);
                 } else {
                     p = new Packet(data, time, eos, streamID, tmpH, sriChanged, false);
-                    this.stats.update(data.length, this.workQueue.size()/this.maxQueueDepth, eos, streamID, false);
+                    this.stats.update(data.length, this.workQueue.size()/(float)this.maxQueueDepth, eos, streamID, false);
                 }
 		if ( logger != null ) {
 		    logger.trace( "bulkio::InPort pushPacket NEW Packet (QUEUE=" + workQueue.size() + ")");
