+++ plugins/bulkio/src/bulkio/InUInt64Port.java
@@ -362,7 +362,7 @@ public class InUInt64Port extends BULKIO.jni.dataUlongLongPOA {
             }
 
             synchronized (this.dataBufferLock) {
-                this.stats.update(data.length, this.workQueue.size()/this.maxQueueDepth, eos, streamID, false);
+                this.stats.update(data.length, this.workQueue.size()/(float)this.maxQueueDepth, eos, streamID, false);
                 this.workQueue.add(p);
                 this.dataSem.release();
             }
@@ -387,7 +387,7 @@ public class InUInt64Port extends BULKIO.jni.dataUlongLongPOA {
                     this.stats.update(data.length, 0, eos, streamID, true);
                 } else {
                     p = new Packet(data, time, eos, streamID, tmpH, sriChanged, false);
-                    this.stats.update(data.length, this.workQueue.size()/this.maxQueueDepth, eos, streamID, false);
+                    this.stats.update(data.length, this.workQueue.size()/(float)this.maxQueueDepth, eos, streamID, false);
                 }
 		if ( logger != null ) {
 		    logger.trace( "bulkio::InPort pushPacket NEW Packet (QUEUE=" + workQueue.size() + ")");
